+++ crates/miden-objects/src/transaction/executed_tx.rs
@@ -12,13 +12,12 @@ use super::{
     OutputNotes,
     TransactionArgs,
     TransactionId,
-    TransactionInputs,
     TransactionOutputs,
-    TransactionWitness,
 };
-use crate::account::{AccountCode, PartialAccount};
+use crate::account::PartialAccount;
 use crate::asset::FungibleAsset;
 use crate::block::BlockNumber;
+use crate::transaction::TransactionInputs;
 use crate::utils::serde::{
     ByteReader,
     ByteWriter,
@@ -46,9 +45,6 @@ pub struct ExecutedTransaction {
     tx_inputs: TransactionInputs,
     tx_outputs: TransactionOutputs,
     account_delta: AccountDelta,
-    tx_args: TransactionArgs,
-    foreign_account_code: Vec<AccountCode>,
-    advice_witness: AdviceInputs,
     tx_measurements: TransactionMeasurements,
 }
 
@@ -64,9 +60,6 @@ impl ExecutedTransaction {
         tx_inputs: TransactionInputs,
         tx_outputs: TransactionOutputs,
         account_delta: AccountDelta,
-        tx_args: TransactionArgs,
-        foreign_account_code: Vec<AccountCode>,
-        advice_witness: AdviceInputs,
         tx_measurements: TransactionMeasurements,
     ) -> Self {
         // make sure account IDs are consistent across transaction inputs and outputs
@@ -86,9 +79,6 @@ impl ExecutedTransaction {
             tx_inputs,
             tx_outputs,
             account_delta,
-            tx_args,
-            foreign_account_code,
-            advice_witness,
             tx_measurements,
         }
     }
@@ -138,7 +128,7 @@ impl ExecutedTransaction {
 
     /// Returns a reference to the transaction arguments.
     pub fn tx_args(&self) -> &TransactionArgs {
-        &self.tx_args
+        self.tx_inputs.tx_args()
     }
 
     /// Returns the block header for the block against which the transaction was executed.
@@ -159,7 +149,7 @@ impl ExecutedTransaction {
     /// Returns all the data requested by the VM from the advice provider while executing the
     /// transaction program.
     pub fn advice_witness(&self) -> &AdviceInputs {
-        &self.advice_witness
+        self.tx_inputs.advice_inputs()
     }
 
     /// Returns a reference to the transaction measurements which are the cycle counts for
@@ -174,22 +164,14 @@ impl ExecutedTransaction {
     /// Returns individual components of this transaction.
     pub fn into_parts(
         self,
-    ) -> (AccountDelta, TransactionOutputs, TransactionWitness, TransactionMeasurements) {
-        let tx_witness = TransactionWitness {
-            tx_inputs: self.tx_inputs,
-            tx_args: self.tx_args,
-            foreign_account_code: self.foreign_account_code,
-            advice_witness: self.advice_witness,
-        };
-
-        (self.account_delta, self.tx_outputs, tx_witness, self.tx_measurements)
+    ) -> (TransactionInputs, TransactionOutputs, AccountDelta, TransactionMeasurements) {
+        (self.tx_inputs, self.tx_outputs, self.account_delta, self.tx_measurements)
     }
 }
 
-impl From<ExecutedTransaction> for TransactionWitness {
+impl From<ExecutedTransaction> for TransactionInputs {
     fn from(tx: ExecutedTransaction) -> Self {
-        let (_, _, tx_witness, _) = tx.into_parts();
-        tx_witness
+        tx.tx_inputs
     }
 }
 
@@ -205,9 +187,6 @@ impl Serializable for ExecutedTransaction {
         self.tx_inputs.write_into(target);
         self.tx_outputs.write_into(target);
         self.account_delta.write_into(target);
-        self.tx_args.write_into(target);
-        self.foreign_account_code.write_into(target);
-        self.advice_witness.write_into(target);
         self.tx_measurements.write_into(target);
     }
 }
@@ -217,20 +196,9 @@ impl Deserializable for ExecutedTransaction {
         let tx_inputs = TransactionInputs::read_from(source)?;
         let tx_outputs = TransactionOutputs::read_from(source)?;
         let account_delta = AccountDelta::read_from(source)?;
-        let tx_args = TransactionArgs::read_from(source)?;
-        let foreign_account_code = <Vec<AccountCode>>::read_from(source)?;
-        let advice_witness = AdviceInputs::read_from(source)?;
         let tx_measurements = TransactionMeasurements::read_from(source)?;
 
-        Ok(Self::new(
-            tx_inputs,
-            tx_outputs,
-            account_delta,
-            tx_args,
-            foreign_account_code,
-            advice_witness,
-            tx_measurements,
-        ))
+        Ok(Self::new(tx_inputs, tx_outputs, account_delta, tx_measurements))
     }
 }