+++ crates/miden-objects/src/transaction/mod.rs
@@ -14,7 +14,6 @@ mod transaction_id;
 mod tx_args;
 mod tx_header;
 mod tx_summary;
-mod tx_witness;
 
 pub use executed_tx::{ExecutedTransaction, TransactionMeasurements};
 pub use inputs::{AccountInputs, InputNote, InputNotes, ToInputNoteCommitments, TransactionInputs};
@@ -31,4 +30,3 @@ pub use transaction_id::TransactionId;
 pub use tx_args::{TransactionArgs, TransactionScript};
 pub use tx_header::TransactionHeader;
 pub use tx_summary::TransactionSummary;
-pub use tx_witness::TransactionWitness;