+++ crates/miden-testing/src/mock_chain/chain.rs
@@ -606,8 +606,8 @@ impl MockChain {
     // INPUTS APIS
     // ----------------------------------------------------------------------------------------
 
-    /// Returns a valid [`TransactionInputs`] for the specified entities, executing against a
-    /// specific block number.
+    /// Returns a valid [`TransactionInputs`] for the specified entities, executing against
+    /// a specific block number.
     pub fn get_transaction_inputs_at(
         &self,
         reference_block: BlockNumber,
@@ -668,7 +668,7 @@ impl MockChain {
         let input_notes = InputNotes::new(input_notes)?;
 
         Ok(TransactionInputs::new(
-            account,
+            account.into(),
             ref_block.clone(),
             partial_blockchain,
             input_notes,