+++ crates/miden-testing/tests/lib.rs
@@ -32,7 +32,7 @@ pub fn prove_and_verify_transaction(
 
     let proof_options = ProvingOptions::default();
     let prover = LocalTransactionProver::new(proof_options);
-    let proven_transaction = prover.prove(executed_transaction.into()).unwrap();
+    let proven_transaction = prover.prove(executed_transaction).unwrap();
 
     assert_eq!(proven_transaction.id(), executed_transaction_id);