+++ crates/miden-tx/src/lib.rs
@@ -6,8 +6,6 @@ extern crate alloc;
 #[cfg(feature = "std")]
 extern crate std;
 
-pub use miden_objects::transaction::TransactionInputs;
-
 mod executor;
 pub use executor::{
     DataStore,